chore: refactoring folder structure
