+++ app/funding/funding-category.tsx
@@ -1,12 +1,17 @@
-import { Button } from "../ui/button";
-import { Form, FormControl, FormField, FormItem } from "../ui/form";
+import { Button } from "../../components/ui/button";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+} from "../../components/ui/form";
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "../ui/select";
+} from "../../components/ui/select";
 import { z } from "zod";
 import { Dispatch, SetStateAction } from "react";
 import { useForm } from "react-hook-form";
