+++ app/funding/funding-step.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import { Button } from "../ui/button";
+import { Button } from "../../components/ui/button";
 import { Dispatch, SetStateAction } from "react";
 
 interface FundingStepPageProps {
