+++ app/inspiring-section.tsx
